LayoutTests/media/video-volume-slider-drag-expected.txt


EVENT(canplaythrough)
EVENT(volumechange)
EVENT(volumechange)
EVENT(volumechange)
EVENT(volumechange)
END OF TEST
